TelegramChannelPush/requirements.txt
PySocks
pytz
telethon
requests
PyJWT